+++ .gitignore
@@ -2,4 +2,3 @@
 
 *.exe
 *.app
-*.out
