+++ code/info.md
@@ -1,14 +1,12 @@
 
 
 Makefile created
-Type make in terminal to compile and then ./main.app to run it
+Type make in terminal to compile and then ./main.out to run it
 Type clean to remove all junk files 
 Type make memcheck to check for memory leaks
 Ask Arnav for questions
 
 
-Stack
-To make a new stack do
-Stack varname = Stack();
-The rest just look at stack.cpp it isn't hard or ask me
-Ask Arnav for questions
+Someone needs to test all the testcases
+Someone needs to make the User manual
+Someone needs to comment the code more
